create account to command
