+++ packages/hedera-wallet-snap/packages/snap/src/utils/CryptoUtils.ts
@@ -21,7 +21,7 @@
 import { PublicKey } from '@hashgraph/sdk';
 import { HDNodeWallet, Mnemonic, assertArgument, ethers } from 'ethers';
 import { DEFAULTCOINTYPE } from '../types/constants';
-import {MirrorNftInfo, MirrorTokenInfo} from '../types/hedera';
+import { MirrorNftInfo, MirrorTokenInfo } from '../types/hedera';
 import { FetchResponse, FetchUtils } from './FetchUtils';
 
 export class CryptoUtils {
